+++ CHANGES.md
@@ -12,8 +12,8 @@ https://github.com/UNICORE-EU/uftp-javaclient/issues
 UFTP Client 2.4.0 (released MMM dd, 2026)
 -----------------------------------------
  - new feature: "cp --dry-run" will only check what would be transferred
- - new feature: "info" will also check if a actual UFTP connection
-   can be established
+ - new feature: "info" will also check if an actual UFTP connection
+   can be established, and also optionally run a quick performance check
  - fix: "ls" output had "true" to indicate a directory instead of "d"
  - improvement: more intuitive chunk size as a result of "cp -T" 
  - improvement: UFTP_OPTIONS environment can be used to set internal